feat: display the state of fragrance switch
